src: fix VC++ warning C4244
Changed usage of v8::Value::InterValue (i.e uint64_t) to Int32Value
fixing VC++ warning C4244.

+++ src/node.cc
@@ -1934,7 +1934,7 @@ static void InitGroups(const FunctionCallbackInfo<Value>& args) {
 void Exit(const FunctionCallbackInfo<Value>& args) {
   Environment* env = Environment::GetCurrent(args.GetIsolate());
   HandleScope scope(env->isolate());
-  exit(static_cast<int>(args[0]->IntegerValue()));
+  exit(args[0]->Int32Value());
 }
 
 
@@ -1986,7 +1986,7 @@ void Kill(const FunctionCallbackInfo<Value>& args) {
     return env->ThrowError("Bad argument.");
   }
 
-  int pid = static_cast<int>(args[0]->IntegerValue());
+  int pid = args[0]->Int32Value();
   int sig = args[1]->Int32Value();
   int err = uv_kill(pid, sig);
   args.GetReturnValue().Set(err);
@@ -3523,7 +3523,7 @@ int EmitExit(Environment* env) {
   process_object->Set(env->exiting_string(), True(env->isolate()));
 
   Handle<String> exitCode = env->exit_code_string();
-  int code = static_cast<int>(process_object->Get(exitCode)->IntegerValue());
+  int code = process_object->Get(exitCode)->Int32Value();
 
   Local<Value> args[] = {
     env->exit_string(),
@@ -3533,7 +3533,7 @@ int EmitExit(Environment* env) {
   MakeCallback(env, process_object, "emit", ARRAY_SIZE(args), args);
 
   // Reload exit code, it may be changed by `emit('exit')`
-  return static_cast<int>(process_object->Get(exitCode)->IntegerValue());
+  return process_object->Get(exitCode)->Int32Value();
 }
